+++ lib/src/grids/grid.dart
@@ -442,35 +442,3 @@ abstract mixin class Grid<T> {
   @override
   String toString() => GridImpl.debugString(this);
 }
-
-/// A mixin that augments [Grid] to indicate it supports efficient index access.
-///
-/// This mixin is used to provide hints to algorithms about the internal layout
-/// of the grid, which can be used to optimize performance for certain
-/// operations.
-///
-/// To implement this mixin, you must override [layoutHint] and provide a
-/// cooresponding implementation for [getByIndexUnchecked]. If the layout hint
-/// is [LayoutHint.private], the behavior is undefined.
-mixin EfficientIndexGrid<T> on Grid<T> {
-  /// A hint to algorithms about the internal layout of the grid.
-  ///
-  /// A value of [LayoutHint.private] indicates that the grid is a private
-  /// implementation detail and should not be relied upon by external code.
-  ///
-  /// > [!IMPORTANT]
-  /// > You _must_ override [getByIndexUnchecked] if a known layout is provided.
-  LayoutHint get layoutHint;
-
-  /// Given a grid of `width * height` cells, returns the nth-cell.
-  ///
-  /// This method is based on the internal layout of the grid, and is not
-  /// guaranteed to be efficient for all grid implementations. If the hint is
-  /// [LayoutHint.private], the behavior is undefined.
-  ///
-  /// > [!WARNING]
-  /// > Use this method with caution, as it may lead to undefined behavior if
-  /// > the index is out of bounds. It is recommended to use [get] instead
-  /// > unless writing performance-sensitive code with proper bounds checking.
-  T getByIndexUnchecked(int index);
-}